Add files via upload
These need a little cleanup, the goal is to create a web-based monitoring system with an Arduino Uno R4.  It will have live updates on power quality and status, its a good start but not functional as of 5/6/25

Once the framework is setup, I'm adding sensors and more cts for additional features and measurements. I'm adding a dashboard so I can have and monitor multiple systems at one time.